requirements.txt
fastapi
python-dotenv
Requests
requests_oauthlib
starlette
uvicorn